requirements.txt
APScheduler==3.10.1
bidict==0.22.1
cffi==1.15.1
click==8.1.3
dnspython==1.16.0
dominate==2.7.0
eventlet==0.30.2
Faker==16.6.1
flake8==6.0.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pysmb==1.2.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pyvips==2.2.1
six==1.16.0
SQLAlchemy==2.0.4
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.5
tzlocal==4.2
visitor==0.1.3
Werkzeug==2.2.2
zope.event==4.6
zope.interface==6.0
